rok-ranking-scanner/requirements.txt
androidviewclient==22.5.1
certifi==2022.12.7
charset-normalizer==3.0.1
clipboard==0.0.4
contourpy==1.0.7
culebratester-client==2.0.64
cycler==0.11.0
fonttools==4.38.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.7.0
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.0
Pillow==9.4.0
protobuf==3.20.2
pyparsing==3.0.9
pyperclip==1.8.2
pytesseract==0.3.10
python-dateutil==2.8.2
requests==2.28.2
six==1.16.0
urllib3==1.26.14
